backend/requirements.txt
paho-mqtt
fins-driver